fix: Move column types to their own module and encapsulate more details
